+++ preview_api/lib/src/channel_methods_sender.dart
@@ -87,4 +87,9 @@ class ChannelMethodsSender with Log implements AbstractChannelMethodsSender {
   void terminatePreview() {
     _invokeMonarchChannelMethod(MonarchMethods.terminatePreview);
   }
+
+  @override
+  void willRestartPreview() {
+    _invokeMonarchChannelMethod(MonarchMethods.willRestartPreview);
+  }
 }